requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==5.0.0
flask-limiter==3.5.0
flask-sqlalchemy==3.1.0
pyjwt==2.8.0
cryptography==44.0.2
requests==2.32.3
sqlalchemy==1.4.40
passlib==1.7.4
python-dotenv==1.1.0
gunicorn==23.0.0
redis==5.2.1
celery==5.5.1
click==8.1.3
pycryptodomex==3.22.0
marshmallow==3.21.2
pyjwt[crypto]==2.8.0
bcrypt==4.1.1
itsdangerous==2.2.0
pytz==2024.1
validators==0.22.0
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
